requirements.txt
mysql-connector-python 
openpyxl 
yagmail 
python-dotenv